Create an "edit note" page
This should ultimately serve for both creating new notes as well as
editing existing ones. If a note id is included in the url and data is
provided, the form should be pre-filled with the content of the note
being edited. Otherwise, generate a blank form for creating a new note.