Add Heroku dependencies and update app config
Heroku uses PostgreSQL, but generates a database url environment
variable that SQLAlchemy doesn't like. We just need a simple string
replacement to fix that up.

Also make sure the app still runs offline. Beef up the logic for
handling whether the DATABASE_URL environment variable is set. The
previous code handled this, but now the interpreter will get mad if we
try to call str.replace() on a nonetype environment variable. The
behavior ultimately remains the same though - the app will assume we're
offline and switch over to using the sqlite database if the variable is
not set.